+++ tests/test_manage.py
@@ -4,11 +4,15 @@ import os
 
 from clay.manage import new, make, run, get_settings, get_current
 import pytest
-from shake import execute
+# from shake import execute
+
+
+def get_cwd():
+    return os.path.dirname(os.path.abspath(__file__)) or '.'
 
 
 def test_get_settings():
-    cwd = os.path.dirname(os.path.realpath(__file__))
+    cwd = get_cwd()
     settings = get_settings(cwd, filename='_test.json')
 
     assert settings
@@ -16,22 +20,21 @@ def test_get_settings():
 
 
 def test_get_current():
-    expected = os.path.normpath(os.path.join(
-        os.path.dirname(os.path.realpath(__file__)), '..'))
+    expected = os.getcwd()
     proto = get_current()
-
-    print os.getcwd()
+    
     assert proto.base_dir == expected
 
 
-def test_make():
-    cwd = os.path.dirname(os.path.realpath(__file__))
-    build_dir = os.path.join(cwd, 'build')
-    try:
-        os.remove(build_dir)
-    except OSError:
-        pass
+# def test_make():
+#     cwd = get_cwd()
+#     build_dir = os.path.join(cwd, 'build')
+#     try:
+#         os.remove(build_dir)
+#     except OSError:
+#         pass
     
-    execute('clay', 'make')
+#     execute('clay', ['make', cwd])
+    
+#     assert os.path.isdir(build_dir)
 
-    assert os.path.isdir(build_dir)